Split the source code into model file and training script
